+++ svc/ServerConnectionService.ts
@@ -1,10 +1,22 @@
+/*
+ * This file belongs to Hoist, an application development toolkit
+ * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
+ *
+ * Copyright © 2024 Extremely Heavy Industries Inc.
+ */
 import {BannerModel} from '@xh/hoist/appcontainer/BannerModel';
 import {HoistService, managed, XH} from '@xh/hoist/core';
 import {Icon} from '@xh/hoist/icon';
 import {Timer} from '@xh/hoist/utils/async';
 import {SECONDS} from '@xh/hoist/utils/datetime';
-import {isEmpty} from 'lodash';
 
+/**
+ * Service to provide heartbeat monitoring for connected server.
+ *
+ * In the case of lost/dropped connection, it will display a banner
+ * indicating state of the server's connection
+ *
+ * */
 export class ServerConnectionService extends HoistService {
     static instance: ServerConnectionService;
 
@@ -14,22 +26,20 @@ export class ServerConnectionService extends HoistService {
     private recoveryCount: number = 0;
     private errorCount: number = 0;
 
-    private conf = {
-        errorThreshold: 5,
-        recoveryThreshold: 3,
-        pingInterval: 1
-    };
+    private conf;
 
     private isShowingError: boolean = false;
 
     override async initAsync() {
-        const conf = XH.getConf('xhServerConnectionConfig', {});
-        if (!isEmpty(conf)) {
-            this.conf = conf;
-        }
+        this.conf = XH.getConf('xhServerConnectionConfig', {
+            errorThreshold: 5,
+            recoveryThreshold: 3,
+            checkIntervalSecs: 1
+        });
+
         this.timer = Timer.create({
             runFn: () => this.pingServer(),
-            interval: this.conf.pingInterval * SECONDS
+            interval: this.conf.checkIntervalSecs * SECONDS
         });
     }
 
@@ -42,8 +52,8 @@ export class ServerConnectionService extends HoistService {
         XH.showBanner({
             category: 'xhAppServerHealth',
             message: showError
-                ? 'Server Unavailable: Unable to establish connection with server'
-                : 'Server Available: Reestablished connection with server',
+                ? 'Server Unavailable: Unable to establish connection'
+                : 'Server Available: Reestablished connection',
             icon: showError ? Icon.warning({size: 'lg'}) : Icon.transaction({size: 'lg'}),
             intent: showError ? 'danger' : 'success',
             sortOrder: BannerModel.BANNER_SORTS.SERVER_CONNECTION,
@@ -51,32 +61,33 @@ export class ServerConnectionService extends HoistService {
         });
     }
 
-    private handlePing(requestOk: boolean) {
-        if (requestOk) {
-            this.recoveryCount += 1;
-            this.errorCount = 0;
-            if (this.recoveryCount === this.conf.recoveryThreshold && this.isShowingError) {
-                this.toggleBanner(false);
-            }
-            this.recoveryCount = this.recoveryCount % this.conf.recoveryThreshold;
-        } else {
-            this.recoveryCount = 0;
-            this.errorCount += 1;
-            if (this.errorCount === this.conf.errorThreshold && !this.isShowingError) {
-                this.toggleBanner(true);
-            }
-            this.errorCount = this.errorCount % this.conf.errorThreshold;
+    private handlePingOK() {
+        this.errorCount = 0;
+        if (!this.isShowingError) return;
+
+        this.recoveryCount += 1;
+        if (this.recoveryCount === this.conf.recoveryThreshold) {
+            this.toggleBanner(false);
+        }
+    }
+
+    private handlePingFail() {
+        this.recoveryCount = 0;
+        if (this.isShowingError) return;
+
+        this.errorCount += 1;
+        if (this.errorCount === this.conf.errorThreshold) {
+            this.toggleBanner(true);
         }
     }
 
     private async pingServer() {
         try {
-            let res = await XH.fetch({
-                url: 'ping'
-            });
-            this.handlePing(res.ok);
+            await XH.fetch({url: 'ping'});
+            this.handlePingOK();
         } catch (exception) {
-            this.handlePing(false);
+            this.logError('Failure in ping call', exception);
+            this.handlePingFail();
         }
     }
 }